services/backend/requirements.txt
alembic
bcrypt
fsspec
fastapi
httpie
jinja2
pandas
pandas_ods_reader
passlib
psycopg2-binary
pytest
python-dotenv
python-multipart
python-jose
pyjwt
requests
sqlalchemy
uvicorn
